Better Mini360 model
Got the trimpot model from
https://grabcad.com/library/evm3e-1

Fixed the inductor using a fileted cuboid, matching the
mesurements of the inductor on the mini360. Material
doesn't match, but that level of accuracy isn't required.

Also, the MP2307 is a smaller SO package. Fixed that and
rearranged the passives. The end result is much more
representative.