Soldering helper jig created with pcb-asm-jig-gen
I checked the 2x20 header for a nice fit.

Then generated the jig for the entire board. Saving
the scad file also as the tool is under heavy
development.

Observations of the 3D print:

PCB fits flush inside the lip. Good thing is that
the base flexes if some of the components get a bit
cross.  Even with this helper, the single row bergs
do give some trouble. Some can rock sideways, due to
the sideways wiggle room available for the legs.
Single pins sit well before the board is inserted,
but they also have this tendency to cross, making it
difficult to fit the board in. There are 8 single
pin headers on the AtiVEGA board, which kind of makes
it a torture test. So a better solution is needed.
An obvious one would be to hollow out the base of
the shells, to allow inserting components _whenever_
required rather than in the beginning.

I found that the female headers that I have here
(in real life) for the Pico are taller that the
one in the KiCAD file (which are the standard ones).
Thus, the jig as is can't be used with the board.
Bummer, but lesson learnt. I had suspected something
was off just before printing, and ran some cross
checks - including with this component, but at this
level the margins for error are small. So, the
model will need to be fixed with some real world
measurements and fixed again.