Fix height of female headers used with Pico
The 1x20 female headers available here locally are taller than the
the 3d model in the design. The plastic shroud is 8.4 mm tall, as
opposed to 7 mm in PinHeader_1x20_P2.54mm_Vertical

To correct this, I have moved the headers and the pico upward up
1.4 mm.

An alternative was the change the model, but I found this simpler
for now. While this exposes the pins of the header in the render,
thus making it look a bit ugly, the actual hardware implementation
is clearer.